Fixed tray and column indexing issues.
This is a fairly chunky fix, summarised as follows:
I've been silly and forgotten to re-index layers that have an index field when others in their collection have been deleted, this means everything goes quickly out of sync.
To fix I have now made it so that each LowerLayer must have an index which is sorted by in the database (and placed into the local structure in the correct order), as such the index field has been removed from each layer (hence index is no longer settable and just returns what indexInParent used to return). When layers are saved to the database the index within its parent is saved as its database index.